+++ shared/wallets/send-form/index.stories.js
@@ -5,7 +5,8 @@ import assetInput, {props3 as assetInputProps} from './asset-input/index.stories
 import chooseAsset from './choose-asset/index.stories'
 import footers from './footer/index.stories'
 import noteAndMemo from './note-and-memo/index.stories'
-import participants from './participants/index.stories'
+import participants, {participantProviderProperties} from './participants/index.stories'
+
 import SendForm from '.'
 
 // TODO some of the state of these child components
@@ -22,6 +23,7 @@ const provider = Sb.createPropProviderWithCommon({
   Header: props => ({}),
   NoteAndMemo: props => ({}),
   Participants: props => ({}),
+  ...participantProviderProperties,
 })
 
 const load = () => {
