+++ shared/wallets/send-form/participants/search.js
@@ -1,10 +1,10 @@
 // @flow
 import * as React from 'react'
-import * as Styles from '../../styles'
-import ResultsList from '../../search/results-list/container'
-import UserInput from '../../search/user-input/container'
-import {Box2} from '../../common-adapters'
-import {ParticipantsRow} from '../common'
+import * as Styles from '../../../styles'
+import ResultsList from '../../../search/results-list/container'
+import UserInput from '../../../search/user-input/container'
+import {Box2} from '../../../common-adapters'
+import {ParticipantsRow} from '../../common'
 
 export type Props = {|
   onClick: (username: string) => void,
@@ -18,7 +18,7 @@ const placeholder = 'Search Keybase'
 // TODO: Once UserInput is cleaned up, we may be able to stretch it
 // properly horizontally without wrapping a vertical Box2 around it.
 const Search = (props: Props) => (
-  <Box2 direction="vertical">
+  <React.Fragment>
     <ParticipantsRow heading="To" style={styles.row}>
       <Box2 direction="vertical" fullWidth={true}>
         <UserInput
@@ -36,13 +36,21 @@ const Search = (props: Props) => (
       onClick={props.onClick}
       onShowTracker={props.onShowTracker}
       disableListBuilding={true}
+      style={styles.list}
     />
-  </Box2>
+  </React.Fragment>
 )
 
 const styles = Styles.styleSheetCreate({
   row: {
     paddingBottom: 0,
+  },
+  list: {
+    position: 'absolute',
+    top: 100,
+    zIndex: 4,
+    backgroundColor: Styles.globalColors.white,
+    width: '100%',
   },
 })
